Refactor: Extract inputToSettings to optimizerSetup.ts
